Move path-util.js from db to common.
